Refactoring. Split metadata functions to two files, edit and query. Tests in wikitextutil work now.
