+++ MoinMoin/macro/MetaTable.py
@@ -34,10 +34,10 @@ from MoinMoin.Page import Page
 from MoinMoin.parser.text_moin_wiki import Parser
 from MoinMoin.wikiutil import form_writer
 
+from MoinMoin.metadata.constants import PROPERTIES
 from MoinMoin.metadata.util import url_escape
-from MoinMoin.metadata import (metatable_parseargs, get_metas,
-                               get_properties, PROPERTIES, 
-                               add_matching_redirs)
+from MoinMoin.metadata.query import (metatable_parseargs, get_metas,
+                                     get_properties, add_matching_redirs)
 from MoinMoin.metadata.wikitextutil import format_wikitext
 
 try:
