+++ MoinMoin/metadata/constants.py
@@ -1,0 +1,17 @@
+# -*- coding: utf-8 -*-"
+# This file includes commonly used constants
+SEPARATOR = '-gwikiseparator-'
+
+# Default node attributes that should not be shown
+SPECIAL_ATTRS = ["gwikilabel", "gwikisides", "gwikitooltip", "gwikiskew",
+                 "gwikiorientation", "gwikifillcolor", 'gwikiperipheries',
+                 'gwikishapefile', "gwikishape", "gwikistyle", 
+                 'gwikicategory', 'gwikiURL', 'gwikiimage', 'gwikiinlinks',
+                 'gwikicoordinates']
+
+CATEGORY_KEY = "gwikicategory"
+TEMPLATE_KEY = "gwikitemplate"
+
+PROPERTIES = ['constraint', 'description', 'hint', 'hidden', 'default']
+
+NO_TYPE = u'_notype'
